refactor gui
