+++ .idea/uiDesigner.xml
@@ -125,5 +125,10 @@
         <default-constraints vsize-policy="3" hsize-policy="3" anchor="0" fill="0" />
       </item>
     </group>
+    <group name="MainPage">
+      <item class="uni.climatemonitor.graphics.MainPage" icon="" removable="true" auto-create-binding="false" can-attach-label="false">
+        <default-constraints vsize-policy="3" hsize-policy="3" anchor="0" fill="0" />
+      </item>
+    </group>
   </component>
 </project>
